AI-model-finance/requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.30.*
pydantic==2.9.*
pyyaml==6.0.*
polars==1.10.*
pandas-ta==0.3.*
lightgbm==4.5.*
optuna==4.0.*
vectorbt==0.27.*
streamlit==1.39.*
python-dotenv==1.0.*
joblib==1.4.*
scikit-learn==1.5.*